Update ScriptSignatureTest with upickle
